+++ Server/apk.txt
@@ -1,0 +1,1 @@
+package:name='com.uptodown'versionCode='480'versionName='4.80'platformBuildVersionName='13'platformBuildVersionCode='33'compileSdkVersion='33'compileSdkVersionCodename='13'install-location:'auto'sdkVersion:'17'targetSdkVersion:'33'uses-permission:name='android.permission.INTERNET'uses-permission:name='android.permission.ACCESS_NETWORK_STATE'uses-permission:name='android.permission.WRITE_EXTERNAL_STORAGE'uses-permission:name='android.permission.GET_ACCOUNTS'uses-permission:name='android.permission.WAKE_LOCK'uses-permission:name='android.permission.GET_PACKAGE_SIZE'uses-permission:name='android.permission.RECEIVE_BOOT_COMPLETED'uses-permission:name='android.permission.USE_CREDENTIALS'uses-permission:name='android.permission.AUTHENTICATE_ACCOUNTS'uses-permission:name='android.permission.MANAGE_ACCOUNTS'uses-permission:name='android.permission.ACCESS_WIFI_STATE'uses-permission:name='android.permission.REQUEST_INSTALL_PACKAGES'uses-permission:name='android.permission.REQUEST_DELETE_PACKAGES'uses-permission:name='android.permission.MANAGE_EXTERNAL_STORAGE'uses-permission:name='android.permission.QUERY_ALL_PACKAGES'uses-permission:name='android.permission.UPDATE_PACKAGES_WITHOUT_USER_ACTION'uses-permission:name='android.permission.POST_NOTIFICATIONS'uses-permission:name='android.permission.INSTALL_PACKAGES'uses-permission:name='android.permission.DELETE_PACKAGES'uses-permission:name='android.permission.ACCESS_SUPERUSER'uses-permission:name='android.permission.RECORD_AUDIO'uses-permission:name='com.google.android.finsky.permission.BIND_GET_INSTALL_REFERRER_SERVICE'uses-permission:name='android.permission.FOREGROUND_SERVICE'uses-permission:name='com.uptodown.DYNAMIC_RECEIVER_NOT_EXPORTED_PERMISSION'application-label:'UptodownAppStore'application-label-af:'UptodownAppStore'application-label-am:'UptodownAppStore'application-label-ar:'UptodownAppStore'application-label-as:'UptodownAppStore'application-label-az:'UptodownAppStore'application-label-be:'UptodownAppStore'application-label-bg:'UptodownAppStore'application-label-bn:'UptodownAppStore'application-label-bs:'UptodownAppStore'application-label-ca:'UptodownAppStore'application-label-cs:'UptodownAppStore'application-label-cs-CZ:'UptodownAppStore'application-label-da:'UptodownAppStore'application-label-da-DK:'UptodownAppStore'application-label-de:'UptodownAppStore'application-label-el:'UptodownAppStore'application-label-en-AU:'UptodownAppStore'application-label-en-CA:'UptodownAppStore'application-label-en-GB:'UptodownAppStore'application-label-en-IN:'UptodownAppStore'application-label-en-XC:'UptodownAppStore'application-label-es:'UptodownAppStore'application-label-es-419:'UptodownAppStore'application-label-es-US:'UptodownAppStore'application-label-et:'UptodownAppStore'application-label-eu:'UptodownAppStore'application-label-fa:'UptodownAppStore'application-label-fi:'UptodownAppStore'application-label-fi-FI:'UptodownAppStore'application-label-fr:'UptodownAppStore'application-label-fr-CA:'UptodownAppStore'application-label-fr-FR:'UptodownAppStore'application-label-gl:'UptodownAppStore'application-label-gu:'UptodownAppStore'application-label-hi:'UptodownAppStore'application-label-hi-IN:'UptodownAppStore'application-label-hr:'UptodownAppStore'application-label-hu:'UptodownAppStore'application-label-hu-HU:'UptodownAppStore'application-label-hy:'UptodownAppStore'application-label-in:'UptodownAppStore'application-label-in-ID:'UptodownAppStore'application-label-is:'UptodownAppStore'application-label-it:'UptodownAppStore'application-label-it-IT:'UptodownAppStore'application-label-iw:'UptodownAppStore'application-label-iw-IL:'UptodownAppStore'application-label-ja:'UptodownAppStore'application-label-ja-JP:'UptodownAppStore'application-label-ka:'UptodownAppStore'application-label-kk:'UptodownAppStore'application-label-km:'UptodownAppStore'application-label-kn:'UptodownAppStore'application-label-ko:'UptodownAppStore'application-label-ko-KR:'UptodownAppStore'application-label-ky:'UptodownAppStore'application-label-lo:'UptodownAppStore'application-label-lt:'UptodownAppStore'application-label-lv:'UptodownAppStore'application-label-mk:'UptodownAppStore'application-label-ml:'UptodownAppStore'application-label-mn:'UptodownAppStore'application-label-mr:'UptodownAppStore'application-label-ms:'UptodownAppStore'application-label-my:'UptodownAppStore'application-label-nb:'UptodownAppStore'application-label-ne:'UptodownAppStore'application-label-nl:'UptodownAppStore'application-label-nl-NL:'UptodownAppStore'application-label-no-NO:'UptodownAppStore'application-label-or:'UptodownAppStore'application-label-pa:'UptodownAppStore'application-label-pl:'UptodownAppStore'application-label-pl-PL:'UptodownAppStore'application-label-pt:'UptodownAppStore'application-label-pt-BR:'UptodownAppStore'application-label-pt-PT:'UptodownAppStore'application-label-ro:'UptodownAppStore'application-label-ru:'UptodownAppStore'application-label-ru-RU:'UptodownAppStore'application-label-si:'UptodownAppStore'application-label-sk:'UptodownAppStore'application-label-sl:'UptodownAppStore'application-label-sq:'UptodownAppStore'application-label-sr:'UptodownAppStore'application-label-sr-Latn:'UptodownAppStore'application-label-sv:'UptodownAppStore'application-label-sv-SE:'UptodownAppStore'application-label-sw:'UptodownAppStore'application-label-ta:'UptodownAppStore'application-label-te:'UptodownAppStore'application-label-th:'UptodownAppStore'application-label-tl:'UptodownAppStore'application-label-tr:'UptodownAppStore'application-label-tr-TR:'UptodownAppStore'application-label-uk:'UptodownAppStore'application-label-ur:'UptodownAppStore'application-label-uz:'UptodownAppStore'application-label-vi:'UptodownAppStore'application-label-zh-CN:'UptodownAppStore'application-label-zh-HK:'UptodownAppStore'application-label-zh-TW:'UptodownAppStore'application-label-zu:'UptodownAppStore'application-icon-120:'res/jV.xml'application-icon-160:'res/jV.xml'application-icon-240:'res/jV.xml'application-icon-320:'res/jV.xml'application-icon-480:'res/jV.xml'application-icon-640:'res/jV.xml'application-icon-65534:'res/jV.xml'application-icon-65535:'res/jV.xml'application:label='UptodownAppStore'icon='res/jV.xml'banner='res/6W.png'launchable-activity:name='com.uptodown.activities.MainActivityScrollable'label='UptodownAppStore'icon=''leanback-launchable-activity:name='com.uptodown.tv.ui.activity.TvMainActivity'label='UptodownAppStore'icon=''banner=''uses-library-not-required:'androidx.window.extensions'uses-library-not-required:'androidx.window.sidecar'uses-permission:name='android.permission.READ_EXTERNAL_STORAGE'uses-implied-permission:name='android.permission.READ_EXTERNAL_STORAGE'reason='requestedWRITE_EXTERNAL_STORAGE'feature-group:label=''uses-feature-not-required:name='android.hardware.camera'uses-feature-not-required:name='android.hardware.faketouch'uses-feature-not-required:name='android.hardware.location.gps'uses-feature-not-required:name='android.hardware.microphone'uses-feature-not-required:name='android.hardware.sensor'uses-feature-not-required:name='android.hardware.touchscreen'uses-feature-not-required:name='android.hardware.wifi'uses-feature-not-required:name='android.software.leanback'provides-component:'search'mainother-activitiesother-receiversother-servicessupports-screens:'small''normal''large''xlarge'supports-any-density:'true'locales:'--_--''af''am''ar''as''az''be''bg''bn''bs''ca''cs''cs-CZ''da''da-DK''de''el''en-AU''en-CA''en-GB''en-IN''en-XC''es''es-419''es-US''et''eu''fa''fi''fi-FI''fr''fr-CA''fr-FR''gl''gu''hi''hi-IN''hr''hu''hu-HU''hy''in''in-ID''is''it''it-IT''iw''iw-IL''ja''ja-JP''ka''kk''km''kn''ko''ko-KR''ky''lo''lt''lv''mk''ml''mn''mr''ms''my''nb''ne''nl''nl-NL''no-NO''or''pa''pl''pl-PL''pt''pt-BR''pt-PT''ro''ru''ru-RU''si''sk''sl''sq''sr''sr-Latn''sv''sv-SE''sw''ta''te''th''tl''tr''tr-TR''uk''ur''uz''vi''zh-CN''zh-HK''zh-TW''zu'densities:'120''160''240''320''480''640''65534''65535'
