--- src/.env
@@ -1,2 +1,0 @@
-REACT_APP_AUTH0_DOMAIN=dev-tumjlmbcbam83fd5.us.auth0.com
-REACT_APP_AUTH0_CLIENT_ID=kX8pQGaSE09v6KurOZmcAMJ20WUhv4Zd
