--- src/apk.txt
@@ -1,1 +1,0 @@
-package:name='me.ccrama.redditslide'versionCode='337'versionName='6.7.1'platformBuildVersionName='10'platformBuildVersionCode='29'compileSdkVersion='29'compileSdkVersionCodename='10'sdkVersion:'16'targetSdkVersion:'29'uses-permission:name='android.permission.INTERNET'uses-permission:name='android.permission.NFC'uses-permission:name='android.permission.WRITE_EXTERNAL_STORAGE'uses-permission:name='com.android.vending.BILLING'uses-permission:name='android.permission.ACCESS_NETWORK_STATE'uses-permission:name='android.permission.ACCESS_WIFI_STATE'uses-permission:name='android.permission.READ_EXTERNAL_STORAGE'application-label:'Slide'application-label-af:'Slide'application-label-af-ZA:'Slide'application-label-am:'Slide'application-label-ar:'Slide'application-label-ar-SA:'Slide'application-label-as:'Slide'application-label-az:'Slide'application-label-be:'Slide'application-label-bg:'Slide'application-label-bn:'Slide'application-label-bs:'Slide'application-label-ca:'Slide'application-label-ca-ES:'Slide'application-label-cs:'Slide'application-label-cs-CZ:'Slide'application-label-da:'Slide'application-label-da-DK:'Slide'application-label-de:'Slide'application-label-de-DE:'Slide'application-label-el:'Slide'application-label-el-GR:'Slide'application-label-en-AU:'Slide'application-label-en-CA:'Slide'application-label-en-GB:'Slide'application-label-en-IN:'Slide'application-label-en-XC:'Slide'application-label-eo-UY:'Slide'application-label-es:'Slide'application-label-es-419:'Slide'application-label-es-ES:'Slide'application-label-es-GT:'Slide'application-label-es-US:'Slide'application-label-et:'Slide'application-label-eu:'Slide'application-label-fa:'Slide'application-label-fi:'Slide'application-label-fi-FI:'Slide'application-label-fil-PH:'Slide'application-label-fr:'Slide'application-label-fr-CA:'Slide'application-label-fr-FR:'Slide'application-label-ga-IE:'Slide'application-label-gl:'Slide'application-label-gl-ES:'Slide'application-label-gu:'Slide'application-label-he:'Slide'application-label-hi:'Slide'application-label-hr:'Slide'application-label-hr-HR:'Slide'application-label-hu:'Slide'application-label-hu-HU:'Slide'application-label-hy:'Slide'application-label-id:'Slide'application-label-in:'Slide'application-label-in-ID:'Slide'application-label-is:'Slide'application-label-it:'Slide'application-label-it-IT:'Slide'application-label-iw:'Slide'application-label-iw-IL:'Slide'application-label-ja:'Slide'application-label-ja-JP:'Slide'application-label-ka:'Slide'application-label-kk:'Slide'application-label-km:'Slide'application-label-kn:'Slide'application-label-ko:'Slide'application-label-ky:'Slide'application-label-lo:'Slide'application-label-lt:'Slide'application-label-lv:'Slide'application-label-lv-LV:'Slide'application-label-mk:'Slide'application-label-ml:'Slide'application-label-mn:'Slide'application-label-mr:'Slide'application-label-ms:'Slide'application-label-my:'Slide'application-label-nb:'Slide'application-label-ne:'Slide'application-label-nl:'Slide'application-label-nl-NL:'Slide'application-label-no-NO:'Slide'application-label-or:'Slide'application-label-pa:'Slide'application-label-pl:'Slide'application-label-pl-PL:'Slide'application-label-pt:'Slide'application-label-pt-BR:'Slide'application-label-pt-PT:'Slide'application-label-ro:'Slide'application-label-ro-RO:'Slide'application-label-ru:'Slide'application-label-ru-RU:'Slide'application-label-si:'Slide'application-label-sk:'Slide'application-label-sl:'Slide'application-label-sq:'Slide'application-label-sq-AL:'Slide'application-label-sr:'Slide'application-label-sr-Latn:'Slide'application-label-sr-SP:'Slide'application-label-sv:'Slide'application-label-sv-SE:'Slide'application-label-sw:'Slide'application-label-ta:'Slide'application-label-te:'Slide'application-label-th:'Slide'application-label-tl:'Slide'application-label-tr:'Slide'application-label-tr-TR:'Slide'application-label-uk:'Slide'application-label-uk-UA:'Slide'application-label-ur:'Slide'application-label-uz:'Slide'application-label-vi:'Slide'application-label-vi-VN:'Slide'application-label-zh:'Slide'application-label-zh-CN:'Slide'application-label-zh-HK:'Slide'application-label-zh-TW:'Slide'application-label-zu:'Slide'application-icon-120:'res/uF.xml'application-icon-160:'res/uF.xml'application-icon-240:'res/uF.xml'application-icon-320:'res/uF.xml'application-icon-480:'res/uF.xml'application-icon-640:'res/uF.xml'application-icon-65534:'res/uF.xml'application:label='Slide'icon='res/uF.xml'launchable-activity:name='me.ccrama.redditslide.Activities.Slide'label='Slide'icon='res/uF.xml'uses-library-not-required:'com.sec.android.app.multiwindow'feature-group:label=''uses-feature:name='android.hardware.faketouch'uses-implied-feature:name='android.hardware.faketouch'reason='defaultfeatureforallapps'uses-feature:name='android.hardware.wifi'uses-implied-feature:name='android.hardware.wifi'reason='requestedandroid.permission.ACCESS_WIFI_STATEpermission'provides-component:'app-widget'provides-component:'notification-listener'mainother-activitiesother-receiversother-servicessupports-screens:'small''normal''large''xlarge'supports-any-density:'true'locales:'--_--''af''af-ZA''am''ar''ar-SA''as''az''be''bg''bn''bs''ca''ca-ES''cs''cs-CZ''da''da-DK''de''de-DE''el''el-GR''en-AU''en-CA''en-GB''en-IN''en-XC''eo-UY''es''es-419''es-ES''es-GT''es-US''et''eu''fa''fi''fi-FI''fil-PH''fr''fr-CA''fr-FR''ga-IE''gl''gl-ES''gu''he''hi''hr''hr-HR''hu''hu-HU''hy''id''in''in-ID''is''it''it-IT''iw''iw-IL''ja''ja-JP''ka''kk''km''kn''ko''ky''lo''lt''lv''lv-LV''mk''ml''mn''mr''ms''my''nb''ne''nl''nl-NL''no-NO''or''pa''pl''pl-PL''pt''pt-BR''pt-PT''ro''ro-RO''ru''ru-RU''si''sk''sl''sq''sq-AL''sr''sr-Latn''sr-SP''sv''sv-SE''sw''ta''te''th''tl''tr''tr-TR''uk''uk-UA''ur''uz''vi''vi-VN''zh''zh-CN''zh-HK''zh-TW''zu'densities:'120''160''240''320''480''640''65534'
